backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
python-dotenv==1.0.0
Werkzeug==2.3.7
reportlab==4.0.4
pandas==2.2.0
pymongo==4.5.0
bcrypt==4.0.1
requests>=2.31.0
numpy>=1.26.0
python-multipart>=0.0.9
jq>=1.6.0
typer>=0.9.0
